galeria

+++ server/api/images/[id].delete.ts
@@ -4,6 +4,10 @@ export default defineEventHandler(async (event) => {
 
   const id = event.context.params?.id
 
+  if (!id) {
+    throw createError({ statusCode: 400, statusMessage: 'ID no válido' })
+  }
+
   try {
     await prisma.image.delete({
       where: { id }
